+++ includes/vm.h
@@ -6,7 +6,7 @@
 /*   By: upopee <upopee@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/05 01:58:14 by upopee            #+#    #+#             */
-/*   Updated: 2018/04/29 04:57:11 by upopee           ###   ########.fr       */
+/*   Updated: 2018/05/02 01:18:59 by upopee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,13 +53,13 @@ void		debug_pcmove(uint32_t pc, uint8_t *memory, uint8_t nb_bytes);
 */
 
 void		debug_infos(t_cwvm *vm, t_vcpu *c, t_gamectrl *g, t_vmverb *v);
-void		debug_memory(uint8_t *arena, uint8_t *a_flags, char *win);
+void		debug_memory(uint8_t *arena, uint16_t *a_flags, char *win);
 void		debug_processes(t_cwvm *vm, t_list *curr, t_jobctrl *jobs);
 void		debug_registers(t_vmverb *verb, t_list *curr);
 
 int			is_pc_val(uint32_t to_test, t_list *p_stack);
 
-uint8_t		get_pno(uint8_t a_flags);
+uint8_t		get_pno(uint16_t a_flags);
 char		*get_timercolor(uint16_t timer);
 char		*get_p_color(uint8_t player_no);
 char		*get_p_bgcolor(uint8_t player_no);
